Project/requirements.txt
gunicorn
django
python-dotenv
whitenoise
cryptography
pyngrok
django-sslserver
qrcode[pil]
djangorestframework